% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/January_PDO.R
\docType{data}
\name{January_PDO}
\alias{January_PDO}
\title{Pacific Decadal Oscillation in January}
\format{
A data frame with 104 rows and 2 variables
}
\usage{
data(January_PDO)
}
\description{
A dataset containing January PDO values. Subset of the data
from NOAA: <https://www.ncei.noaa.gov/access/monitoring/pdo/>.
}
\details{
\itemize{
  \item PDO - Pacific Decadal Oscillation in January for the given year.
  \item Age - the year for which the PDO was measured.
}
}
\keyword{datasets}
