% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_data.R
\docType{data}
\name{lake_data}
\alias{lake_data}
\title{DCA-ordinated pollen data from Lake Consuelo}
\format{
A data frame with 39 rows and 2 variables
}
\usage{
data(lake_data)
}
\description{
A dataset containing pre-processed DCA-ordinated data from 
Bush, Silman & Urrego (2004) <doi:10.1126/science.1090795>.
}
\details{
\itemize{
  \item DCA1 - DCA values for each timepoint from the raw dataset.
  \item Age - timepoint of each sample that has been DCA-ordinated.
}
}
\keyword{datasets}
