% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_RSI.R
\docType{data}
\name{lake_RSI}
\alias{lake_RSI}
\title{DCA-ordinated pollen data from Lake Consuelo with RSI values}
\format{
A data frame with 39 rows and 3 variables
}
\usage{
data(lake_RSI)
}
\description{
A dataset containing pre-processed DCA-ordinated data from 
Bush, Silman & Urrego (2004) <doi:10.1126/science.1090795>.
This data has been processed using Rodionov(lake_data, "DCA1", "Age", l=5, merge=TRUE)
}
\details{
\itemize{
  \item DCA1 - DCA values for each timepoint from the raw dataset.
  \item Age - timepoint of each sample that has been DCA-ordinated.
  \item RSI - Regime Shift Index (see docs for Rodionov()) for each timepoint.
}
}
\keyword{datasets}
