% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regime_means.R
\name{regime_means}
\alias{regime_means}
\title{Calculate means for each regime}
\usage{
regime_means(data, col, rsi)
}
\arguments{
\item{data}{The dataframe that will be used.}

\item{col}{The column we are measuring change on.}

\item{rsi}{The column containing RSI values.}
}
\value{
A vector of the mean value for each regime.
}
\description{
calculates the mean for each regime in a regime shift analysis.
}
\examples{
regime_means(lake_RSI, "DCA1", "RSI")

}
