% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{rshift-package}
\alias{rshift}
\alias{rshift-package}
\title{rshift: Paleoecology Functions for Regime Shift Analysis}
\description{
Contains a variety of functions, based around regime shift analysis of paleoecological data. Citations: Rodionov() from Rodionov (2004) \doi{10.1029/2004GL019448} Lanzante() from Lanzante (1996) \doi{10.1002/(SICI)1097-0088(199611)16:11\%3C1197::AID-JOC89\%3E3.0.CO;2-L} Hellinger_trans from Numerical Ecology, Legendre & Legendre (ISBN 9780444538680) rolling_autoc from Liu, Gao & Wang (2018) \doi{10.1016/j.scitotenv.2018.06.276} Sample data sets lake_data & lake_RSI processed from Bush, Silman & Urrego (2004) \doi{10.1126/science.1090795} Sample data set January_PDO from NOAA: \url{https://www.ncei.noaa.gov/access/monitoring/pdo/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexhroom/rshift}
  \item Report bugs at \url{https://github.com/alexhroom/rshift/issues}
}

}
\author{
\strong{Maintainer}: Alex H. Room \email{alex.room@btinternet.com} (\href{https://orcid.org/0000-0002-5314-2331}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Felipe Franco-Gaviria (\href{https://orcid.org/0000-0003-4799-1457}{ORCID}) [contributor, funder]
  \item Dunia H. Urrego (\href{https://orcid.org/0000-0001-7938-5529}{ORCID}) [contributor, funder]
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
