% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_regime_means}
\alias{rust_regime_means}
\title{Calculates the mean for each regime in a regime shift analysis.}
\usage{
rust_regime_means(col, rsi)
}
\arguments{
\item{col}{The column we are measuring change on.}

\item{rsi}{The column containing RSI values.}
}
\description{
Calculates the mean for each regime in a regime shift analysis.
}
