% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_rodionov}
\alias{rust_rodionov}
\title{Calculate STARS RSI points and return to R as a vector}
\usage{
rust_rodionov(vals, t_crit, l)
}
\description{
Calculate STARS RSI points and return to R as a vector
}
\keyword{internal}
