% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_rodionov_huber}
\alias{rust_rodionov_huber}
\title{Calculate STARS RSI points and return to R as a vector
Uses Huber weighting to handle outliers}
\usage{
rust_rodionov_huber(vals, t_crit, l, huber)
}
\arguments{
\item{t_crit}{The critical value of a t-distribution at the desired p-value}

\item{l}{The cut-off length of a regime; affects sensitivity}

\item{huber}{The tuning constant for Huber weighting.}

\item{input_vals}{The column we are measuring change on}
}
\description{
Calculate STARS RSI points and return to R as a vector
Uses Huber weighting to handle outliers
}
