% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_to_percentage.R
\name{absolute_to_percentage}
\alias{absolute_to_percentage}
\title{Converts absolute abundance data to a percentage of total abundance for each site}
\usage{
absolute_to_percentage(data, col, site)
}
\arguments{
\item{data}{The dataframe to be used.}

\item{col}{The column that change is being measured on.}

\item{site}{The column containing the site of each sample.}
}
\value{
The `data` dataframe with an added 'percentage' column.
}
\description{
Converts absolute abundance data to a percentage of total abundance for each site
}
