% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_autoc.R
\name{rolling_autoc}
\alias{rolling_autoc}
\title{Rolling autocorrelation}
\usage{
rolling_autoc(data, col, l)
}
\arguments{
\item{data}{The dataframe that will be used.}

\item{col}{The column we are measuring change on.}

\item{l}{The time interval (no. of columns) used in the autocorrelation.}
}
\value{
A table of rolling lag-1 autocorrelation values.
}
\description{
finds lag-1 autocorrelation in a rolling window; can be used to predict resilience (Liu, Gao, & Wang, 2018)
}
