% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSI_graph.R
\name{RSI_graph}
\alias{RSI_graph}
\title{Regime Shift Index graph}
\usage{
RSI_graph(data, col, time, rsi)
}
\arguments{
\item{data}{The dataframe that will be used.}

\item{col}{The column we are measuring change on.}

\item{time}{The column containing time units (e.g. age of a subsample)}

\item{rsi}{The column containing RSI values (for best visualisation use Rodionov() with merge=TRUE)}
}
\value{
Two graphs, one on top of the other; one of col against time and one of RSI against time.
}
\description{
creates two graphs, one of data and one of the RSI, as seen in Rodionov (2004)
}
\examples{
RSI_graph(lake_RSI, "DCA1", "Age", "RSI")
}
